Comments
In the old code, I noticed it showed a red error at the very beginning of the code, 1 to be exact. So I left the 1st part blank, and started removing the 2nd part of the code. Aside from the fact that I noticed that the Hello Worlds changed to Hello Vi, I noticed that some of the structures remained the same.